# docs/05-api/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory API
  version: 1.0.0
  description: API interna y pública para gestión de inventario StockFlow.
  contact:
    name: StockFlow Support
    email: soporte@tuempresa.com
servers:
  - url: https://inventario.tuempresa.com
    description: Production
  - url: http://localhost:8080
    description: Local backend
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login de usuario
      description: Autentica un usuario con credenciales (email/password) y devuelve los tokens JWT de acceso y refresco.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: Solicitud mal formada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refrescar token de acceso
      description: Genera un nuevo Access Token JWT utilizando la cookie httpOnly que contiene el Refresh Token.
      operationId: refreshToken
      responses:
        '200':
          description: Token renovado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: Nuevo token JWT de acceso.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhZG1pbkB2ZGVuZXJneS5lcyIsImV4cCI6MTgwMDB9...
        '401':
          description: Token de refresco ausente o inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Cerrar sesión
      description: Revoca el token de refresco actual y limpia las credenciales de la sesión.
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Cierre de sesión exitoso
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/materials:
    get:
      tags: [Materials]
      summary: Listar e investigar materiales de inventario
      description: Devuelve una lista de materiales según los filtros opcionales aplicados.
      operationId: listMaterials
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filtrar por estado físico.
          required: false
          schema:
            type: string
            enum: [OPERATIVO, ROTO, EN_REPARACION, BAJA]
        - name: officePublicId
          in: query
          description: Filtrar por identificador de oficina asociada.
          required: false
          schema:
            type: string
        - name: materialType
          in: query
          description: Filtrar por tipo o categoría de equipo.
          required: false
          schema:
            type: string
        - name: serialNumber
          in: query
          description: Filtrar por coincidencia parcial de número de serie.
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
      responses:
        '200':
          description: Materiales encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageMaterialResponse'
    post:
      tags: [Materials]
      summary: Registrar nuevo material
      description: Da de alta un nuevo artículo, le asocia una oficina y auto-genera su código público y código QR único de inventario.
      operationId: createMaterial
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialCreateRequest'
      responses:
        '201':
          description: Material creado y registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/materials/{publicCode}:
    get:
      tags: [Materials]
      summary: Consultar detalle de un material
      description: Obtiene la información técnica completa, estado y trazabilidad física del material por su código público.
      operationId: getMaterial
      security:
        - bearerAuth: []
      parameters:
        - name: publicCode
          in: path
          required: true
          description: Código único seguro de inventario (mat_XXXXXXXX).
          schema:
            type: string
      responses:
        '200':
          description: Detalle del material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
        '404':
          description: Material no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      tags: [Materials]
      summary: Modificar información de material
      description: Modifica los parámetros descriptivos, número de serie o asignación de sede de un equipo.
      operationId: updateMaterial
      security:
        - bearerAuth: []
      parameters:
        - name: publicCode
          in: path
          required: true
          description: Código único seguro de inventario (mat_XXXXXXXX).
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialCreateRequest'
      responses:
        '200':
          description: Datos del material actualizados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
...
  /api/v1/materials/{publicCode}/history:
    get:
      tags: [Inventory]
      summary: Obtener historial de movimientos del material
      description: Recupera cronológicamente todos los traslados de oficina, transiciones de estado y comentarios justificativos vinculados al material.
      operationId: getMaterialHistory
      security:
        - bearerAuth: []
      parameters:
        - name: publicCode
          in: path
          required: true
          description: Código único seguro de inventario (mat_XXXXXXXX).
          schema:
            type: string
      responses:
        '200':
          description: Lista de historial obtenida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaterialHistoryResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Introducir el Access Token JWT precedido del prefijo 'Bearer '.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clave de acceso estática para integradores externos autorizados.
  parameters:
    pageParam:
      name: page
      in: query
      description: Número de página a consultar (base 0).
      required: false
      schema:
        type: integer
        default: 0
    sizeParam:
      name: size
      in: query
      description: Cantidad máxima de registros por página.
      required: false
      schema:
        type: integer
        default: 10
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: Dirección de correo institucional del usuario.
          example: admin@tuempresa.com
        password:
          type: string
          format: password
          description: Contraseña de la cuenta.
          example: password123
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Token de autenticación JWT para llamadas API subsiguientes.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: Token de refresco seguro de larga duración (enviado también en Cookie).
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.ref...
        tokenType:
          type: string
          description: Formato de autorización requerido.
          example: Bearer
        expiresIn:
          type: integer
          description: Tiempo de expiración del access token en segundos.
          example: 900
...
    MaterialResponse:
      type: object
      properties:
        publicCode:
          type: string
          example: mat_a1b2c3d4
        materialType:
          type: string
          example: Router
        brand:
          type: string
          example: MikroTik
        model:
          type: string
          example: RB4011
        serialNumber:
          type: string
          example: SN_MIKRO_98765
        description:
          type: string
          example: Router troncal de comunicación principal
        status:
          type: string
          enum: [OPERATIVO, ROTO, EN_REPARACION, BAJA]
          example: OPERATIVO
        officePublicId:
          type: string
          example: off-c8d7e6f5
        officeName:
          type: string
          example: Sede Central Madrid
        createdByName:
          type: string
          example: Juan Pérez
        updatedByName:
          type: string
          example: Juan Pérez
        createdAt:
          type: string
          format: date-time
          example: "2026-05-22T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-22T08:30:00Z"
    MaterialCreateRequest:
      type: object
      required: [materialType, brand, model, serialNumber, officePublicId]
      properties:
        materialType:
          type: string
          example: Switch
        brand:
          type: string
          example: Cisco
        model:
          type: string
          example: Catalyst 2960L
        serialNumber:
          type: string
          example: SN_CISCO_445566
        description:
          type: string
          example: Switch PoE de 24 puertos
        officePublicId:
          type: string
          example: off-c8d7e6f5
...
    MaterialHistoryResponse:
      type: object
      properties:
        id:
          type: integer
          example: 124
        materialPublicCode:
          type: string
          example: mat_a1b2c3d4
        statusBefore:
          type: string
          example: OPERATIVO
        statusAfter:
          type: string
          example: BAJA
        officeNameBefore:
          type: string
          example: Sede Central Madrid
        officeNameAfter:
          type: string
          example: Sede Central Madrid
        comment:
          type: string
          example: Retirado por obsolescencia tecnológica.
        performedByName:
          type: string
          example: Juan Pérez
        createdAt:
          type: string
          format: date-time
          example: "2026-05-22T08:30:00Z"
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: La justificación es obligatoria para dar de baja un equipo.
        timestamp:
          type: string
          format: date-time
          example: "2026-05-22T08:30:01Z"
...
    PageMaterialResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MaterialResponse'
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 25
        totalPages:
          type: integer
          example: 3
        last:
          type: boolean
          example: false
...